requirements.txt
amqp==5.1.0
asgiref==3.3.2
attrs==21.4.0
autopep8==1.5.4
billiard==3.6.4.0
black==22.1.0
cachetools==5.0.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.2.1
cryptography==36.0.2
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.7
django-cors-headers==3.5.0
django-countries==7.2.1
django-environ==0.8.1
django-filter==21.1
django-phonenumber-field==5.2.0
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-rest-passwordreset==1.2.1
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
factory-boy==3.2.1
Faker==8.9.1
flake8==4.0.1
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.4
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
phonenumbers==8.12.33
pika==1.1.0
Pillow==8.3.2
platformdirs==2.5.1
pluggy==1.0.0
prompt-toolkit==3.0.28
protobuf==3.19.4
psycopg2-binary==2.9.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.1.1
pytest-cov==2.12.1
pytest-django==4.4.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==5.3.1
requests==2.27.1
requests-oauthlib==1.3.1
retry==0.9.2
rsa==4.8
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.1
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
django-rest-swagger==2.2.0
graphviz==0.19.2
django-extensions==3.1.5
pydot ==1.4.2
